backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
azure-storage-blob==12.24.0
python-multipart==0.0.20
gunicorn==21.2.0
azure-identity==1.14.0

pandas
openpyxl
